Extract deterministic AOI scoring into pick_aoi/scoring.py
